+++ src/engineering.rs
@@ -34,8 +34,8 @@ use crate::coordinates::Coordinate;
 use crate::math::{RigidBodyTransform, Rotation};
 use crate::systems::EquivalentTo;
 use crate::{Point3, Vector};
-use std::marker::PhantomData;
-use std::ops::Mul;
+use core::marker::PhantomData;
+use core::ops::Mul;
 use uom::si::f64::{Angle, Length};
 use uom::ConstZero;
 
@@ -905,6 +905,7 @@ mod tests {
     use crate::coordinates::Coordinate;
     use crate::directions::{Bearing, Components as BearingComponents};
     use crate::engineering::{Orientation, Pose};
+    use crate::float_math::FloatMath;
     use crate::geodetic::{Components as Wgs84Components, Wgs84};
     use crate::math::{RigidBodyTransform, Rotation};
     use crate::util::BoundedAngle;
@@ -913,6 +914,7 @@ mod tests {
     use approx::assert_relative_eq;
     use approx::{assert_abs_diff_eq, AbsDiffEq};
     use rstest::rstest;
+    use std::println;
     use uom::si::f64::{Angle, Length};
     use uom::si::{
         angle::{degree, radian},
@@ -1127,7 +1129,7 @@ mod tests {
     #[case(
         Point3::new(10., 10., -10.),
         (d(0.), d(0.), d(0.)),
-        Bearing::build(BearingComponents { azimuth: d(45.), elevation: r((10. / (10_f64.powi(2) * 3.).sqrt()).asin()) }).unwrap()
+        Bearing::build(BearingComponents { azimuth: d(45.), elevation: r(FloatMath::asin(10. / FloatMath::sqrt(FloatMath::powi(10_f64, 2) * 3.))) }).unwrap()
     )]
     fn pose_direction_towards(
         #[case] position: Point3,