+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support `no_std` environments via new `std` (default) and `libm`
+  feature flags ([#59](https://github.com/helsing-ai/sguaba/pull/59))
+
 ### Changed
 
 ### Deprecated